Update compatibility of Client class
